Rename namespace to Latest
