+++ src/Microsoft.Performance.SDK.Runtime/DTO/Latest/ColumnConfiguration.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Runtime.Serialization;
 
-namespace Microsoft.Performance.SDK.Runtime.DTO.V1_3
+namespace Microsoft.Performance.SDK.Runtime.DTO.Latest
 {
     [DataContract]
     internal class ColumnConfiguration
