+++ src/Microsoft.Performance.SDK.Runtime/DTO/Latest/TableConfigurations.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Runtime.Serialization;
 
-namespace Microsoft.Performance.SDK.Runtime.DTO.V1_3
+namespace Microsoft.Performance.SDK.Runtime.DTO.Latest
 {
     [DataContract]
     internal class TableConfigurations
