+++ src/Microsoft.Performance.SDK.Runtime/DTO/V1_0/ColumnConfiguration.cs
@@ -7,7 +7,7 @@ namespace Microsoft.Performance.SDK.Runtime.DTO.V1_0
 {
     [DataContract]
     internal class ColumnConfiguration
-        : ISupportUpgrade<V1_3.ColumnConfiguration>
+        : ISupportUpgrade<Latest.ColumnConfiguration>
     {
         /// <summary>
         ///     Metadata describing the column.
@@ -21,9 +21,9 @@ namespace Microsoft.Performance.SDK.Runtime.DTO.V1_0
         [DataMember]
         public UIHints DisplayHints { get; set;  }
 
-        public V1_3.ColumnConfiguration Upgrade()
+        public Latest.ColumnConfiguration Upgrade()
         {
-            return new V1_3.ColumnConfiguration()
+            return new Latest.ColumnConfiguration()
             {
                 Metadata = this.Metadata,
                 VariantGuid = null,
