+++ src/Microsoft.Performance.SDK.Runtime/DTO/V1_0/PrebuiltConfigurations.cs
@@ -8,7 +8,7 @@ namespace Microsoft.Performance.SDK.Runtime.DTO.V1_0
 {
     [DataContract]
     internal class PrebuiltConfigurations
-        : PreviousPrebuiltConfigurationBase<V1_3.PrebuiltConfigurations>
+        : PreviousPrebuiltConfigurationBase<Latest.PrebuiltConfigurations>
     {
         internal static readonly double DTOVersion = 1.0;
 
@@ -20,9 +20,9 @@ namespace Microsoft.Performance.SDK.Runtime.DTO.V1_0
         [DataMember(Order = 2)]
         public TableConfigurations[] Tables { get; set; }
 
-        protected override V1_3.PrebuiltConfigurations UpgradeToNext()
+        protected override Latest.PrebuiltConfigurations UpgradeToNext()
         {
-            return new V1_3.PrebuiltConfigurations()
+            return new Latest.PrebuiltConfigurations()
             {
                 Tables = this.Tables.Select(configs => configs.Upgrade()).ToArray()
             };
